+++ src/commands/backups/raw/create.py
@@ -291,6 +291,9 @@ def main(  # noqa: CFQ002 (max arguments)
     dry_run: bool = typer.Option(
         False, "--dry-run", "-n", help="Do not actually backup, only show what would be done."
     ),
+    skip_root_check: bool = typer.Option(
+        False, "--skip-root-check", help="Do not cancel when not run with root privileges."
+    ),
 ):
     """
     Backup files and directories.
@@ -298,7 +301,7 @@ def main(  # noqa: CFQ002 (max arguments)
 
     obj: BbakContextObject = ctx.obj
 
-    if not (dry_run or os.geteuid() == 0):
+    if not skip_root_check and not (dry_run or os.geteuid() == 0):
         raise DocoError(
             "You need to have root privileges to create/download/restore a backup.\n"
             "Please try again, this time using 'sudo'."